[TECH-521] Rename the build artifacts module
branch: feature/TECH-521-implement-remote-artifact-caching
